Add Splash Activity feature
