+++ app/src/main/java/com/bluecat/githubfeed/main/MainActivityView.kt
@@ -14,18 +14,8 @@
  * limitations under the License.
  */
 
-package com.bluecat.githubfeed
+package com.bluecat.githubfeed.main
 
-import com.bluecat.core.BasePresenter
-import timber.log.Timber
+import com.bluecat.core.BaseView
 
-class MainPresenter : BasePresenter<MainActivityView>() {
-
-    init {
-        Timber.d("Initialize MainPresenter.")
-    }
-
-    fun getHelloMessage(): String {
-        return "hello, GitHub Feed!"
-    }
-}
+interface MainActivityView : BaseView
