+++ app/src/main/java/com/bluecat/githubfeed/splash/SplashActivityView.kt
@@ -1,0 +1,5 @@
+package com.bluecat.githubfeed.splash
+
+import com.bluecat.core.BaseView
+
+interface SplashActivityView : BaseView
