+++ app/src/main/java/com/bluecat/githubfeed/splash/SplashPresenter.kt
@@ -1,0 +1,7 @@
+package com.bluecat.githubfeed.splash
+
+import com.bluecat.core.BasePresenter
+
+class SplashPresenter:BasePresenter<SplashActivityView>() {
+
+}
